test: change promises to async/await
